backend/projectx/requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anyio==4.11.0
asgiref==3.9.2
attrs==25.4.0
autobahn==25.10.2
Automat==25.4.16
billiard==4.2.2
boto3==1.40.59
botocore==1.40.59
celery==5.5.3
certifi==2025.8.3
cffi==2.0.0
channels==4.3.1
channels_redis==4.3.0
charset-normalizer==3.4.3
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
coinbase-commerce==1.0.1
colorama==0.4.6
constantly==23.10.4
cron_descriptor==2.0.6
cryptography==46.0.3
daphne==4.2.1
distro==1.9.0
dj-database-url==3.0.1
Django==5.2.7
django-celery-beat==2.8.1
django-cors-headers==4.9.0
django-encrypted-model-fields==0.6.5
django-filter==25.1
django-rest-framework==0.1.0
django-storages==1.14.6
django-timezone-field==7.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
jiter==0.11.1
jmespath==1.0.1
kombu==5.5.4
msgpack==1.1.1
numpy==2.3.3
openai==2.8.0
packaging==25.0
pandas==2.3.3
pillow==12.0.0
polygon-api-client==1.16.3
prompt_toolkit==3.0.52
psycopg==3.2.12
psycopg-binary==3.2.12
psycopg-pool==3.2.7
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.3
pydantic_core==2.41.4
PyJWT==2.10.1
pyOpenSSL==25.3.0
pyotp==2.9.0
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.2
redis==5.2.1
requests==2.32.5
s3transfer==0.14.0
service-identity==24.2.0
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
stripe==13.0.0
tqdm==4.67.1
tradingeconomics==4.5.5
Twisted==25.5.0
txaio==25.9.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.14
websocket-client==1.9.0
websockets==15.0.1
whitenoise==6.11.0
zope.interface==8.0.1
